// src/app/shared/types/api-types/bido-equipment-dto.interface.ts
import { BidoEquipmentAttributeDTO } from './bido-equipment-attribute-dto.interface';
import { StockInformationDTO } from './stock-information-dto-interface';

export interface BidoEquipmentDTO {
  equipmentCode?: string;
  familyVariantCode?: string;
  equEquipmentCode?: string;
  ownerCustomerCode?: string;
  pnCode?: string;
  operatorCustomerCode?: string;
  sn?: string;
  snManufacturer?: string;
  equipmentDesignation?: string;
  equipmentFunction?: string;
  manufacturingDate?: Date;
  chapter?: string;
  section?: string;
  subject?: string;
  sheet?: string;
  marks?: string;
  equiPosition?: string;
  sentence?: string;
  manufacturingConfiguration?: boolean;
  functionCode?: string;
  zone?: string;
  country?: string;
  registration?: string;
  eisDate?: Date;
  lifeRestriction?: boolean;
  operationalStatus?: string;
  modelCode?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  bidoEquipmentAttributeList?: BidoEquipmentAttributeDTO[];
  batchNumber?: string;
  stockInformationDTO?: StockInformationDTO;
  batchType?: string;
  quantity?: number;
  logisticsSequenceCode?: string;
  _displayableField?: string;
}