// src/app/shared/types/api-types/bidt-stock-mvt-dto.interface.ts
import { BidoEquipmentDTO } from './bido-equipment-dto.interface';
import { BidtSiteDTO } from './bidt-site-dto.interface';
import { BidtStockMvtAttributeDTO } from './bidt-stock-mvt-attribute-dto.interface';
import { BidtStockMvtOperationDTO } from './bidt-stock-mvt-operation-dto.interface';
import { BidtStockMvtTypeDTO } from './bidt-stock-mvt-type-dto.interface';
import { BidtStorageBinDTO } from './bidt-storage-bin-dto.interface';
import { BidtWarehouseDTO } from './bidt-warehouse-dto.interface';

export interface BidtStockMvtDTO {
  bidtSalesRequestId?: number;
  bidtStockMvtId?: number;
  bidtExternalUserId?: number;
  bidtValuationGroupId?: number;
  bidtPurchaseRequestId?: number;
  bidoNotificationNotificationCode?: string;
  bsdeInventoryProjectId?: string;
  bsdeInventoryInvId?: string;
  bidtStockTypeId?: number;
  bidmWorkOrderProjectId?: string;
  bidmWorkOrderWoId?: string;
  birePnPnCode?: string;
  bidtStorageBinId?: number;
  bidtWarehouseId?: number;
  stockMvtCode?: string;
  stockMvtName?: string;
  stockMvtWay?: string;
  stockMvtQuantity?: string;
  mvtStatus?: string;
  stockStatus?: string;
  stockMvtDate?: Date;
  stockMvtSn?: string;
  stockMvtBatchNumber?: string;
  stockMvtComment?: string;
  stockOwnerCode?: string;
  stockIsExternal?: number;
  counterCode?: string;
  numberOfDays?: number;
  priceValue?: string;
  priceUnit?: string;
  byQuotation?: number;
  discountValue?: string;
  discountPercentValue?: string;
  statusState?: string;
  statusUser?: string;
  statusDate?: Date;
  id?: number;
  bidtWarehouseByBidtWarehouseReceiptId?: number;
  bidtDfItemId?: number;
  bidtStorageBinByBidtSbBinReceiptId?: number;
  bidtWarehouseByBidtWarehouseIssueId?: number;
  bidtWarehouseByBidtWarehouseId?: number;
  bidtStorageBinByBidtSbZoneIssueId?: number;
  bidtStorageBinByBidtSbBinIssueId?: number;
  bidtStorageBinByBidtSbZoneReceiptId?: number;
  bidtTorItemId?: number;
  bidtSiteByBidtSiteReceiptId?: number;
  bidtStockMvtTypeId?: number;
  bidtProcReqItemId?: number;
  bidtSiteByBidtSiteIssueId?: number;
  bidtStorageBinByBidtStorageBinId?: number;
  requester?: string;
  requestedPotentialDate?: Date;
  requestedPotentialValue?: string;
  requestedPotentialUnit?: string;
  priorityLevel?: string;
  priorityReason?: string;
  stockMvtPlannedDate?: Date;
  criticality?: string;
  pending?: number;

  bidtStockMvtAttributeDTOs?: BidtStockMvtAttributeDTO[];
  bidtStockMvtOperations?: BidtStockMvtOperationDTO[];

  bidtStockMvtType?: BidtStockMvtTypeDTO;
  bidtWarehouseIssue?: BidtWarehouseDTO;
  bidtWarehouseReceipt?: BidtWarehouseDTO;
  bidtSiteIssue?: BidtSiteDTO;
  bidtSiteReceipt?: BidtSiteDTO;
  bidtStorageBinBinIssue?: BidtStorageBinDTO;
  bidtStorageBinBinReceipt?: BidtStorageBinDTO;
  bidtStorageBinZoneIssue?: BidtStorageBinDTO;
  bidtStorageBinZoneReceipt?: BidtStorageBinDTO;

  bidoEquipmentDTO?: BidoEquipmentDTO;
  birePnArticleDesignation?: string;
  equipmentOperationalStatus?: string;
  stockMvtUnit?: string;
}